requirements.txt
python-telegram-bot==20.0a2
Telethon~=1.24.0
python-dotenv~=0.20.0